FIX: Fix bug in getUserById(int userId, PostRaf postRaf , UserRelationRaf userRelationRaf)
Problem: Return an unexpected and wrong result.
Solution: Because of small scale of project, I moved searching user from file to memory. replaced raf search with stream. It looks more concise and neat.

Note: solution may not be proper with more number of user, as it consume a lot of memory.